debian-packaging: Fix missing tilde case in lexical comparisons
Back in #11, I derived the overall structure of the resulting `match`
block from combining the tilde / end-of-string comparisons from the
former one + the tilde alphanumeric comparisons from `compare_char`.
Unfortunately, in the process, an important case got omitted: comparing
any character on the left to a tilde on the right, which was incorrectly
replaced with comparing to end-of-string on the left.

This restores the correct match branch, as well as regrouping the
comparisons and adding extra comments for readability.

Closes #27.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>